requirements.txt
django-appconf
